+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+* `Gillespie` no longer has a const parameter.
+
 ## [0.3.1] -- 2021-08-01
 
 Increased performance by 3-5% by optimizing exponential sampling.